contracts-bedrock: fix `OptimismMintableERC20Created`
The names of the event args were backwards from what they
were supposed to be. I noticed this while writing integration
tests for the sdk.

The event is used like this:

```solidity
emit OptimismMintableERC20Created(_remoteToken, address(localToken), msg.sender);
```

The event was updated to look like this:

```solidity
event OptimismMintableERC20Created(
    address indexed remoteToken,
    address indexed localToken,
    address deployer
);
```

The `remoteToken` is first and the `localToken` is second.