op-node: Cleanup calldata source API
This provides a new set of objects which provide the following API guarantes:
- The opening of a data for a new L1 block will never.
- If it failed to get transactions, it does so in the calls to `Next`.

This greatly simplifies usage of this object when constructing new data. If
the node did not keep trakc of the internal state, the external users of this
API would have to keep track of this state in a more complex way.